aiq: Release v4.0x9.8-rc0 for RV1106 beta SDK
Summary:
 - Support write AWB debug buffer to file
   driver version should match with kernel
 - Support switch sub-scene
 - Support cutomer ae/awb
 - Optimized code size
 - Optimized memory usages
 - Optimized cpu ultilization
 - Fix params bug for 3dlut/ccm/gic..

Updated CLs:
07228c12 aiq: prepare v4.0x9.8-rc0 release
2d6c79a2 Adrc: Fix drc enable switch bug in Tool stManual
c8ef505f AE：simplify translator codes
c6878a3d Add working mode info to translator
be0179d2 AE: simplify codes
f2bc244a aiq: iqfiles: sc3336/sc530ai: Fix json object name wrong
6b7ac17f awb : 1 delete some memset operation in PrepareMeasureResult
ea5075e1 j2s: modify MAX_NAME from 64 to 32
8b822ef6 aiq: cac: Reduce json object name size < 32
4d736a37 aiq: ispp: Reduce json object name size < 32
d9ce7466 abayer2dnrV23 abayertnrV23 ynrv22 cnrV30 gainV2 sharpV33:
e94f6707 CCM/3DLUT: Reduce structure name size
c2e298f0 AF: change enum type size
fca59b4d AE: decrease length of enum
54f81b36 awb : reduce the length of type to 32 jin j2s
c150a03a iqfiles:add sc3336 & sc530ai json
37d322f5 aiq: Disable build RkAiqCorePpThread
6f84cef1 aiq: Do not build pdaf related code if it is not enabled
80e427d1 aiq: Fix build error if luma core diabled
03877c0f awb : support third party
614263a3 aiq: build: Do not build AF by default
e5842ae5 aiq: Fix build error on IPC SDK
6095e4b5 awb : support to write input data for awb stat
08845020 aiq: Support alloc dbg info buffer for awb&gain
2e541220 kernel headers: Add support awb dbg info to ddr
b33beac9 Condition compilation for RkLumaCore
3391695f Condition compilation for IspXXXParams
c4f4d828 j2s: disable cache file
2c059623 support fragment json iq.
cbb609ed Tuning: fix memory leak step 2.
aa1d4616 aiq: core: Fix build error if thumbnail disabled
30a38baa 3dlut: fix bug in manual cfg
...


Former-commit-id: bc30c2eb9e1fb59ad4adcc3c6e9bca4ce0fcdac0

+++ rkaiq/include/algos/agic/rk_aiq_types_algo_agic_int.h
@@ -140,8 +140,8 @@ typedef struct rkaiq_gic_v2_param_selected_s {
     uint16_t k_grad2_dark;
     uint16_t min_grad_thr_dark1;
     uint16_t min_grad_thr_dark2;
-    float NoiseScale;
-    float NoiseBase;
+    int NoiseScale;
+    int NoiseBase;
     float noiseCurve_0;
     float noiseCurve_1;
     float globalStrength;
@@ -239,8 +239,8 @@ typedef struct AgicConfigV21_s {
     unsigned short regmingradthr2;
     unsigned short regmingradthr1;
     unsigned char gr_ratio;
-    float noise_scale;
-    float noise_base;
+    uint8_t noise_scale;
+    uint16_t noise_base;
     float noiseCurve_0;
     float noiseCurve_1;
     float sigma_y[15];
